+++ ReplaceableMusicMod/Utilities/ScreenshotGenerator.cs
@@ -6,18 +6,18 @@ namespace MusicReplacer.Utilities;
 
 public static class ScreenshotGenerator
 {
-    public static void GenerateScreenshotWithDelay(int triggerId, float delay)
+    public static void GenerateScreenshotWithDelay(long triggerId, float delay)
     {
         Plugin.StartCoroutineOnPlugin(GenerateScreenshotWithDelayCoroutine(triggerId, delay));
     }
     
-    private static IEnumerator GenerateScreenshotWithDelayCoroutine(int triggerId, float delay)
+    private static IEnumerator GenerateScreenshotWithDelayCoroutine(long triggerId, float delay)
     {
         yield return new WaitForSeconds(delay);
         GenerateScreenshotForTrigger(triggerId);
     }
     
-    public static void GenerateScreenshotForTrigger(int triggerId)
+    public static void GenerateScreenshotForTrigger(long triggerId)
     {
         var level = GameController.Instance.GetCurrentLevelDefinition().levelName;
         
@@ -29,12 +29,12 @@ public static class ScreenshotGenerator
         ScreenCapture.CaptureScreenshot(path);
     }
 
-    public static string GetScreenshotFilePath(string levelName, int triggerId)
+    public static string GetScreenshotFilePath(string levelName, long triggerId)
     {
         return Path.Combine(FileManagement.GetScreenshotsFolder(), GetScreenshotFileName(levelName, triggerId));
     }
     
-    private static string GetScreenshotFileName(string levelName, int triggerId)
+    private static string GetScreenshotFileName(string levelName, long triggerId)
     {
         return $"{levelName}_{triggerId}.png";
     }
